refactor: moved utility for onboarding
