+++ src/payloads/features/idpay/onboarding/types.ts
@@ -35,10 +35,5 @@ export enum IDPayInitiativeID {
   ERR_CHECK_BUDGET_TERMINATED,
   ERR_CHECK_ENDED,
   ERR_CHECK_NOT_STARTED,
-  ERR_CHECK_SUSPENDED,
-
-  NOT_CONFIGURED,
-  CONFIGURED,
-  UNSUBSCRIBED,
-  SUSPENDED
+  ERR_CHECK_SUSPENDED
 }
