+++ src/payloads/features/idpay/timeline/data.ts
@@ -34,7 +34,6 @@ import { CardInfo } from "../../../../../generated/definitions/pagopa/CardInfo";
 import { WalletV2 } from "../../../../../generated/definitions/pagopa/WalletV2";
 import { getWalletV2 } from "../../../../routers/walletsV2";
 import { ibanList } from "../iban/data";
-import { IDPayInitiativeID as InitiativeId } from "../types";
 import { RefundDetailDTO } from "../../../../../generated/definitions/idpay/RefundDetailDTO";
 
 const wallet: WalletV2 = getWalletV2()[0];
